test-requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-socket==0.6.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
requests==2.31.0
aiohttp==3.9.1
python-socketio==5.9.0
python-engineio==4.7.1
flask==2.3.3
flask-socketio==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
azure-cosmos==4.5.1
python-dotenv==1.0.0
